execution/rlp: fold Addr test into decode_test.go
